requirements.txt
geopy~=2.4.1
folium~=0.18.0
python-dotenv~=0.9.1